+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Changed
 - Bootstrap load as a local asset (was loaded from CDN)
 - tests to PHPUnit 9
+- composer/installers support to latest v1 and added v2
 
 ## 0.5 - 2019-11-21